feat: remove option to print bibliography template
It turns out that, as pointed in [1], it is not currently possible
to pass paths to the template function, as they will not be taken
into account when rendering the template. This means that we
have to either:

1. Come up with a way to pass the bibliography content to the
dissertation template, to be rendered there. This would involve
most likely saving the return of #bibliography into a parameter.
2. Make the user print the bibliography themselves.

Considering that, once this is accepted in typst universe, the
users will start from a pre-populated template, I thought best
to follow step 1.