feat: re-introduce the package module
Before having our lib.typ re-exporting the public interface of
the template, I decided to remove the package.typ from the
src because having it in tum-tastic-thesis meant that the
package object also had to be re-exported to the users. This
is no longer the case, so we can centralize our external packages
in one file again.